+++ src/client/widgets/sliders/encoder.js
@@ -44,7 +44,7 @@ module.exports = class Encoder extends StaticProperties(Knob, {angle: 360, range
             }
         })
 
-        defaults.scripting.script.help.push(
+        defaults.scripting.onValue.help.push(
             'Additionnal variables:',
             '- `locals.speed`: encoder\'s speed (reduce `sensitivity` to increase averaging)',
             '- `locals.angle`: encoder\'s angle in degrees'